moved tests to inner module
Otherwise skews coverage info
